+++ Src/Private/Add-DiaNodeImage.ps1
@@ -1,12 +1,12 @@
 function Add-DiaNodeImage {
     <#
     .SYNOPSIS
-        Generates an HTML table for visualizing an icon with customizable properties, such as border, style, and image size.
+        Generates an HTML table for visualizing an image with customizable properties, such as border, style, and image size.
 
     .DESCRIPTION
-        The Add-DiaNodeImage function creates an HTML table to display an icon image, typically used for diagramming nodes.
-        It supports customization of the icon's appearance, including border width, color, style, and image size percentage.
-        The function also allows for debug mode, which highlights the table for easier troubleshooting, and supports specifying the icon image via a hashtable object.
+        The Add-DiaNodeImage function creates an HTML table to display an image, typically used for diagramming nodes.
+        It supports customization of the image's appearance, including border width, color, style, and image size percentage.
+        The function also allows for debug mode, which highlights the table for easier troubleshooting, and supports specifying the image via a hashtable object.
 
     .PARAMETER Name
         Specifies the name of the node to be illustrated. This is a required parameter.
@@ -47,7 +47,7 @@ function Add-DiaNodeImage {
 
     .NOTES
         Author: Jonathan Colon
-        Version: 0.2.30
+        Version: 0.2.34
         Twitter: @jcolonfzenpr
         Github: rebelinux