+++ Src/Private/ConvertTo-Pdf-WaterMark.ps1
@@ -5,7 +5,7 @@ function ConvertTo-Pdf-WaterMark {
     .DESCRIPTION
         Export a diagram in PDF/PNG/SVG formats using PSgraph.
     .NOTES
-        Version:        0.2.31
+        Version:        0.2.34
         Author:         Jonathan Colon
         Bluesky:        @jcolonfpr.bsky.social
         Github:         rebelinux
@@ -35,18 +35,32 @@ function ConvertTo-Pdf-WaterMark {
             HelpMessage = 'Please provide the path to the image output file'
         )]
         [ValidateScript({
-            $parentPath = Split-Path -Path $_ -Parent
-            if (-not ($parentPath | Test-Path) ) {
-                throw "Folder does not exist"
-            }
-            return $true
-        })]
+                $parentPath = Split-Path -Path $_ -Parent
+                if (-not ($parentPath | Test-Path) ) {
+                    throw "Folder does not exist"
+                }
+                return $true
+            })]
         [String]$DestinationPath
     )
     process {
         try {
             Write-Verbose -Message "Trying to convert $($ImageInput.Name) object to PDF format. Destination Path: $DestinationPath."
-            & $ImageMagickPath -quality 100 $ImageInput.FullName $DestinationPath
+            switch ($PSVersionTable.PSEdition) {
+                'Core' {
+                    # Net 9.0 assembly call
+                    [Diagrammer.ConvertImageToPDF]::ConvertPngToPdf($ImageInput.FullName, $DestinationPath)
+                }
+                'Desktop' {
+                    # Net 4.8 assembly call
+                    [DiaConvertImageToPDF.ConvertImageToPDF]::ConvertPngToPdf($ImageInput.FullName, $DestinationPath)
+                }
+                Default {
+                    # Net 4.8 assembly call (Fucking shit)
+                    [DiaConvertImageToPDF.ConvertImageToPDF]::ConvertPngToPdf($ImageInput.FullName, $DestinationPath)
+                }
+            }
+
         } catch {
             Write-Verbose -Message "Unable to convert $($ImageInput.Name) object to PDF format."
             Write-Debug -Message $($_.Exception.Message)